requirements.txt
Flask
marko
gunicorn
google-cloud-storage
firebase-admin
PIL
